refactor: move docs search logic into own hook
