+++ packages/common/hooks/index.ts
@@ -2,6 +2,7 @@ export * from './useBreakpoint'
 export * from './useConstant'
 export * from './useCopy'
 export * from './useDebounce'
+export * from './useDocsSearch'
 export * from './useParams'
 export * from './useTelemetryProps'
 export * from './useThemeSandbox'
